Move `CodableHackrNewsFeedStore` to production
