+++ HackrNewsTests/Stories Cache/CodableHackrNewsFeedStoreTests.swift
@@ -5,72 +5,6 @@
 import HackrNews
 import XCTest
 
-class CodableHackrNewsFeedStore: HackrNewsFeedStore {
-    private struct Cache: Codable {
-        let news: [CodableHackrNew]
-        let timestamp: Date
-
-        var localFeed: [LocalHackrNew] {
-            news.map(\.local)
-        }
-    }
-
-    private struct CodableHackrNew: Codable {
-        private let id: Int
-
-        init(_ localHackrNew: LocalHackrNew) {
-            id = localHackrNew.id
-        }
-
-        var local: LocalHackrNew {
-            LocalHackrNew(id: id)
-        }
-    }
-
-    private let storeUrl: URL
-
-    init(storeURL: URL) {
-        storeUrl = storeURL
-    }
-
-    func retrieve(completion: @escaping RetrievalCompletion) {
-        guard let data = try? Data(contentsOf: storeUrl) else {
-            return completion(.empty)
-        }
-        do {
-            let decoder = JSONDecoder()
-            let decoded = try decoder.decode(Cache.self, from: data)
-            completion(.found(news: decoded.localFeed, timestamp: decoded.timestamp))
-        } catch {
-            completion(.failure(error))
-        }
-    }
-
-    func insertCacheNews(_ news: [LocalHackrNew], with timestamp: Date, completion: @escaping InsertionCompletion) {
-        do {
-            let encoder = JSONEncoder()
-            let cache = Cache(news: news.map(CodableHackrNew.init), timestamp: timestamp)
-            let encoded = try encoder.encode(cache)
-            try encoded.write(to: storeUrl)
-            completion(nil)
-        } catch {
-            completion(error)
-        }
-    }
-
-    func deleteCachedNews(completion: @escaping DeletionCompletion) {
-        do {
-            guard FileManager.default.fileExists(atPath: storeUrl.path) else {
-                return completion(.none)
-            }
-            try FileManager.default.removeItem(at: storeUrl)
-            completion(.none)
-        } catch {
-            completion(error)
-        }
-    }
-}
-
 final class CodableHackrNewsFeedStoreTests: XCTestCase {
     override func setUp() {
         super.setUp()
